refactor: StringUtils를 util 패키지로 이동
