+++ src/test/java/softeer2nd/chess/util/StringUtilsTest.java
@@ -1,7 +1,8 @@
-package softeer2nd.chess;
+package softeer2nd.chess.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import softeer2nd.chess.util.StringUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
